third_party/blink/web_tests/external/wpt/custom-elements/registries/scoped-custom-element-registry-customelementregistry-attribute-expected.txt

This is a testharness.js-based test.
[FAIL] HTML parser should create a builtin element with null registry if customelementregistry is set
  assert_equals: expected (object) null but got (undefined) undefined
[FAIL] Setting customelementregistry content attribute after a builtin element had finishsed parsing should not set null registry
  assert_equals: expected (object) object "[object CustomElementRegistry]" but got (undefined) undefined
[FAIL] Cloning a builtin element with null regsitry should create an element with null registry
  assert_equals: expected (object) null but got (undefined) undefined
[FAIL] HTML parser should create a custom element candidate with null registry if customelementregistry is set
  assert_equals: expected (object) null but got (undefined) undefined
[FAIL] Setting customelementregistry content attribute after a custom element candidate had finishsed parsing should not set null registry
  assert_equals: expected (object) object "[object CustomElementRegistry]" but got (undefined) undefined
[FAIL] Cloning a custom element candidate with null regsitry should create an element with null registry
  assert_equals: expected (object) null but got (undefined) undefined
[FAIL] HTML parser should create a custom element with null registry if customelementregistry is set
  assert_equals: expected (object) null but got (undefined) undefined
[FAIL] Setting customelementregistry content attribute after a custom element had finishsed parsing should not set null registry
  assert_equals: expected (object) object "[object CustomElementRegistry]" but got (undefined) undefined
[FAIL] Cloning a custom element with null regsitry should create an element with null registry
  assert_equals: expected (object) null but got (undefined) undefined
[FAIL] Descendants of an element with customelementregistry should use null registry
  assert_equals: expected (object) null but got (undefined) undefined
[FAIL] Setting customelementregistry content attribute during constructor should not make it use null registry
  assert_equals: expected (object) object "[object CustomElementRegistry]" but got (undefined) undefined
Harness: the test ran to completion.
